Added new class structure for supporting new Mass and NET alignment architecture
Former-commit-id: 0bcc18943db7f05e5bf5bf4c5e00e43e38436724
Former-commit-id: 0e445036f9b85935b3e236ee8d3049a482cd1923

+++ DeconTools.Backend/Core/ScanNETPair.cs
@@ -3,14 +3,14 @@ namespace DeconTools.Backend.Core
 {
     public class ScanNETPair
     {
-        public ScanNETPair(float scan, float net)
+        public ScanNETPair(double scan, double net)
         {
             Scan = scan;
             NET = net;
         }
 
-        public float Scan {get;set;}
-        public float NET { get; set; }
+        public double Scan {get;set;}
+        public double NET { get; set; }
 
 
         public override string ToString()
